# swagger-specs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ChatGPT API",
    "version": "1.0.0",
    "description": "ChatGPT API文档，包含流式和非流式接口",
    "contact": {
      "name": "开发团队"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "开发服务器"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT认证，在请求头中添加：Authorization: Bearer {token}"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "tags": [
          "认证"
        ],
        "summary": "用户登录/注册",
        "description": "使用用户名/邮箱和密码登录，如果用户不存在则自动注册新账号，获取JWT访问令牌和刷新令牌",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "login",
                  "password"
                ],
                "properties": {
                  "login": {
                    "type": "string",
                    "description": "用户名或电子邮件"
                  },
                  "password": {
                    "type": "string",
                    "format": "password"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功或注册并登录成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "accessToken": {
                      "type": "string"
                    },
                    "refreshToken": {
                      "type": "string"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "username": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        },
                        "role": {
                          "type": "string"
                        }
                      }
                    },
                    "isNewUser": {
                      "type": "boolean",
                      "description": "指示是否为新注册用户"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误"
          },
          "401": {
            "description": "认证失败"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "tags": [
          "认证"
        ],
        "summary": "刷新访问令牌",
        "description": "使用刷新令牌获取新的访问令牌，刷新令牌在Authorization请求头中提供",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "刷新成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "accessToken": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "无效的刷新令牌"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": [
          "认证"
        ],
        "summary": "用户登出",
        "description": "使当前访问令牌和刷新令牌失效",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "登出成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未认证"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
    "/api/auth/validate": {
      "get": {
        "tags": [
          "认证"
        ],
        "summary": "验证令牌",
        "description": "验证当前请求中的令牌是否有效",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "令牌有效",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "valid": {
                      "type": "boolean"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "username": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        },
                        "role": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "令牌无效"
          }
        }
      }
    },
...
  },
...
}